requirements.txt
bcrypt~=4.0.1
msgpack~=1.0.5
PyYAML~=6.0
requests~=2.31.0
PyQt5~=5.15.9
qt-material~=2.14
wireguard-tools~=0.4.1
psycopg~=3.1.9
Flask~=2.3.2